requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
django-config==0.1.4
django-crispy-forms==1.9.1
django-phone-field==1.8.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.4
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
